Message: add ability to filter notifiable messages (within a certain range)
Added support for finding messages that fall within a certain offset range.
This functionality is required by Enrollment to identify which messages are
safe to queue up to send to the messaging hub.

The notifiable scope takes an offset which should represent how far along
an enrollment is, in days.  Assuming an offset of X, the scope will identify
messages that fall between X-1 and X+5 (inclusive) days from stream start.