Seed Data: support for loading Message and MessageStream data from YML files
Although the notifier will be able to pull message and message stream data
from the hub eventually, we need the ability to import that data from a
local source; it is possible that the hub might be unavailable due to poor
network connectivity or the notifier is being run locally for development
purposes.  An external repo has been created called 'hms-data' that will
allow the hub and the notifier to share the same data source for seed data.

The hms-data repo is currently private until licensing can be confirmed.

$> git clone git@github.com:dhedlund/hms-data.git db/seed_data
$> rake db:seed