Enrollment: allow successive enrollments, prevent concurrent enrollments
Previous database schema was designed to prevent concurrent enrollments
based on the phone number and message stream.  This prevented the use case
of someone enrolling, cancelling or finishing and then trying to enroll
again in the same stream at a later date.  This limitation became obvious
when considering what happens on successive pregnancies.